requirements.txt
gradio>=4.0.0
torch>=2.0.1
transformers>=4.30.2
diffusers>=0.21.4
accelerate>=0.20.3
numpy>=1.24.3
scikit-learn>=1.3.0
pillow>=10.0.0
pyyaml>=6.0
safetensors>=0.3.3